+++ test/parser.js
@@ -6,13 +6,13 @@ var Parser = require('../lib/parser');
 describe('Parse', function() {
     [ '', 'a', 'a^2', 'a^2+b^{2}', 'l_a^2+l_b^2=l_c^2' ].forEach(function(e) {
         it('should parse: '+JSON.stringify(e), function() {
-            var p = Parser.parse(e);
+            var p = Parser.parse(e).result;
             //console.log(JSON.stringify(e), '->', p.toString());
         });
     });
     it('should parse texvc example', function() {
         var e = '\\sin(x)+{}{}\\cos(x)^2 newcommand';
-        var p = Parser.parse(e);
+        var p = Parser.parse(e).result;
         //console.log(JSON.stringify(e), '->', p.toString());
     });
 });