+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+# texvcjs x.x.x (not yet released)
+
 # texvcjs 0.2.0 (2014-07-23)
 
 * Bug fixes to \begin{...} ... \end{...} environment parsing.