+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Changed return value of low-level `parse` API to allow providing
   more detailed information about the TeX functions encountered.
+* Switch to text mode before emitting \AA and \textvisiblespace.
 
 # texvcjs 0.2.0 (2014-07-23)