+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# texvcjs x.x.x (not yet released)
+# texvcjs 0.3.0 (2014-07-24)
 
 * Switch to text mode before emitting \AA and \textvisiblespace.
 * Require mandatory argument for \overbrace, \overleftarrow,